Documents/smoke-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
playwright==1.40.0
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.1
